Create Create_Training_DATA.py
This Code converts our image samples and convert them into Training Data for our Neural Network. In this code you convert image to an array of data so neural network can understand it. We also resize our data as we don't need high resolution image for this program and small resolution images are faster and easier to work with. Then we also apply labels to our image which is very important and store it in array. We need to covert it to Numpy array to use with neural network. Lastly we use pickle to save our data so we can use it later.